reverting the changes from the changeset
